Split BotMessage into two components
- Refactored structure and logic for how conditional rendering happens inside BotMessage
- Now a PlaceholderMessage is first shown, which is then replaced by a FinishedBotMessage
